Update C99 datatypes

+++ src/gmenu2x.h
@@ -92,7 +92,7 @@ struct MenuOption {
 	MenuAction action;
 };
 
-char *ms2hms(unsigned long t, bool mm, bool ss);
+char *ms2hms(uint32_t t, bool mm, bool ss);
 
 class Menu;
 
@@ -135,8 +135,8 @@ private:
 
 #ifdef TARGET_GP2X
 	typedef struct {
-		unsigned short batt;
-		unsigned short remocon;
+		uint16_t batt;
+		uint16_t remocon;
 	} MMSP2ADC;
 
 	int batteryHandle;
@@ -146,7 +146,7 @@ private:
 		usbnet,
 		samba,
 		web;
-	volatile unsigned short *MEM_REG;
+	volatile uint16_t *MEM_REG;
 	int cx25874; //tv-out
 	void gp2x_tvout_on(bool pal);
 	void gp2x_tvout_off();
@@ -175,9 +175,9 @@ public:
 	/*
 	 * Variables needed for elements disposition
 	 */
-	uint resX, resY, halfX, halfY;
-	// uint bottomBarIconY, bottomBarTextY
-	uint linkColumns, linkRows;
+	uint32_t resX, resY, halfX, halfY;
+	// uint32_t bottomBarIconY, bottomBarTextY
+	uint32_t linkColumns, linkRows;
 	SDL_Rect listRect, linksRect, sectionBarRect;
 	/*!
 	Retrieves the parent directory of GMenu2X.
@@ -221,11 +221,11 @@ public:
 	Reads the current battery state and returns a number representing it's level of charge
 	@return A number representing battery charge. 0 means fully discharged. 5 means fully charged. 6 represents a gp2x using AC power.
 	*/
-	unsigned short getBatteryLevel();
-	long getBatteryStatus();
+	uint16_t getBatteryLevel();
+	int32_t getBatteryStatus();
 
 	void skinMenu();
-	uint onChangeSkin();
+	uint32_t onChangeSkin();
 	void initLayout();
 
 	bool inputCommonActions(bool &inputAction);
@@ -263,7 +263,7 @@ public:
 	void contextMenu();
 	void changeWallpaper();
 
-	void setCPU(unsigned int mhz);
+	void setCPU(uint32_t mhz);
 	const string getDateTime();
 	void setDateTime();
 
@@ -293,7 +293,7 @@ public:
 	int drawButton(Button *btn, int x=5, int y=-10);
 	int drawButton(Surface *s, const string &btn, const string &text, int x=5, int y=-10);
 	int drawButtonRight(Surface *s, const string &btn, const string &text, int x=5, int y=-10);
-	void drawScrollBar(uint pagesize, uint totalsize, uint pagepos, SDL_Rect scrollRect);
+	void drawScrollBar(uint32_t pagesize, uint32_t totalsize, uint32_t pagepos, SDL_Rect scrollRect);
 
 	Menu* menu;
 };
